Avoid unnecessary sysroot rebuilds
Instead of listing each package's debs in a deparate section,
combine all the lists of debs into a set and then sort it into
a single list.

This means debs are listed once only and the order is stable.
In turn this means the multistrap.conf won't change if the
required packages changed but the required debs didn't.