Add i2c-dev.h
In Stretch and older this file is diverted which causes problems
for multistrap. Adding the header to the overlay means it will
always be available.

In Buster and later the diversion is no longer needed as the
header has been renamed. When Raspbian Buster is released this
workaround can be removed.