Fix up brcm pkgconfig workaround.
The previous workaround in de704578 does a bit more than it should
by removing the pkgconfig dependency links to bcm_host. This library
is actually required. Qt will add it without the help from pkgconfig
but keeping the dependency is still correct.

This means only brcmegl.pc needs to have its contents modified. It
now depends on bcm_host and also explicitly links it, which is
redundant. Linking vcoshostif is still unnecessary.

This patch restores the bcm_host dependencies and moves the patched
.pc files into /opt/vc next to the originals.

Unmodified .pcs are now symlinked instead of overlayed.

The new path for .pc files is added in the build script so that the
configs can be found.

See raspberrypi/firmware#1013